Improved hmxm generator

+++ src/cuda/hmxm/cudaHMXM.cu
@@ -457,7 +457,7 @@ int main( int argc, char* argv[] )
     					#pragma omp critical
     					{
 
-    						snprintf(error_detail, 150, "p: [%d, %d], r: %1.16e, e: %1.16e", i, j, A[i + k * j], GOLD[i + k * j]);
+    						snprintf(error_detail, 150, "p: [%d, %d], r: %1.16e, e: %1.16e", i, j, (float)(A[i + k * j]), (float)(GOLD[i + k * j]));
     						if (verbose && (host_errors < 10)) printf("%s\n", error_detail);
     						#ifdef LOGS
     						log_error_detail(error_detail);
